+++ src/org/team4909/bionicframework/hardware/BionicSpeedController.java
@@ -1,4 +1,4 @@
-package org.team4909.bionicframework.devices;
+package org.team4909.bionicframework.hardware;
 
 import edu.wpi.first.wpilibj.SpeedController;
 import edu.wpi.first.wpilibj.command.Command;
