+++ src/org/team4909/powerup2018/Robot.java
@@ -9,13 +9,13 @@ import edu.wpi.first.wpilibj.Spark;
 import edu.wpi.first.wpilibj.command.Command;
 import edu.wpi.first.wpilibj.smartdashboard.SendableChooser;
 
-import org.team4909.bionicframework.devices.Arduino;
-import org.team4909.bionicframework.devices.BionicSRX;
-import org.team4909.bionicframework.devices.BionicSolenoid;
-import org.team4909.bionicframework.devices.BionicSpeedController;
-import org.team4909.bionicframework.devices.PotentiometerController;
-import org.team4909.bionicframework.devices.RoboRio;
-import org.team4909.bionicframework.devices.Arduino.State;
+import org.team4909.bionicframework.hardware.Arduino;
+import org.team4909.bionicframework.hardware.BionicSRX;
+import org.team4909.bionicframework.hardware.BionicSolenoid;
+import org.team4909.bionicframework.hardware.BionicSpeedController;
+import org.team4909.bionicframework.hardware.PotentiometerController;
+import org.team4909.bionicframework.hardware.RoboRio;
+import org.team4909.bionicframework.hardware.Arduino.State;
 import org.team4909.bionicframework.oi.BionicF310;
 
 public class Robot extends RoboRio {
